Add $rows argument, change behaviour
User can now specify the number of rows to print. It'll print all the
rows if --notes is passed or if $rows evaluates to false or if the
user passed value greater than the length of Array.

It'll now warn the user if file doesn't exist locally & user has
passed --local, also instead of just dying it fetches the file from
the web instead.

$covid_19_data not doesn't get printed if --notes is passed. We assume
that the user either wants to see notes or $covid_19_data.