Use ctime instead of mtime to compare
HTTP::Tiny sets mtime to "Last Modified" time set by the server. ctime
is set as the time when that file gets created on disk so we use that
to compare now, actually this should've been used from the start but
File::Fetch didn't set mtime differently (?) so mtime worked till we
used File::Fetch.