Fix rows option behaviour
The behaviour was broken because previously it wouldn't have printed
as many rows as user had entered because of other options.

For example, say if user ran `ara --rows 2 --hide india`, previously
this would've printed only 1 row because "India" is always the first
row & the foreach loop would've looped only two times so
$rows_to_print was more like number of rows to iterate over.

Now it does what it says, it prints that many number of rows.