Fix file fetching logic, use mtime, switch to HTTP::Simple
The issue with ctime is that it doesn't change & stays the same as
when the file was actually created, this means the code was running
the if block basically everytime after 8 minutes of file creation.

And I couldn't use mtime because HTTP::Tiny's mirror method stores
mtime as 'Last Modified' header.

This switch to HTTP::Simple will download the file everytime which
means the mtime changes & we can make the program not look for file
every time. However this also means that we are downloading the whole
file everytime instead of only when it changes. Meh... It's okay
because the alternative is to write a complicated logic.